Adjusted logic of CubeIntake
Before, if B was pressed and a limit switch was actuated, the motor would be clicking continuously between 0% and mCubeWheelSpeed%. It would probably stay at 0 because of motor friction, but this still isn't optimal. I think you should find my changes to be logically equivalent to what you wanted to happen; if the limit switches are pushed in, they will only allow either movement out or zero movement.