Fixed Wheel Spin Direction
Made the wheels spin the same direction when bringing the Cube out or in, this was due to a miscommunication. Also increased the speed of the wheels as 0.3 is too slow with the chains. Using the max speed of 1.0 may mean that the delay on the autonEjectCube can be reduced as it now takes significantly less time to eject the Cube.