Changed Back To PIDF on Position + Fixed Error Bug + Zero Gyro in Auto
For future reference, there's a bug with getClosedLoopError where it won't update at the proper timestep. This might be due to the synchronization of the drive method, but maybe not.

If we calculate error by ourselves, it works fine though.

Also have to zero gyro in autonomous. Placement gets weird otherwise.