Changed Arm Implementation to a Single Driven Solenoid
Changed to a single driven solenoid due to pneumatic design. I'm not a fan of this design personally, as it effectively removes a number of modes of use (the output in the rotational degree of freedom is now binary which effectively removes the degree of freedom from a controls perspective), but there's nothing I can do about that.